refactor: simplify HeaderAuthActions and extract HeaderSignInModal component
