+++ src/app/_components/HeaderActions.tsx
@@ -1,19 +1,12 @@
 'use client';
 
-import { useCallback, useEffect, useState } from 'react';
+import { useState } from 'react';
 import { FaBars, FaGithub, FaRegLightbulb, FaTelegramPlane } from 'react-icons/fa';
-import HeaderAuthActions from '@/app/_components/HeaderAuthActions';
-import SignInModal from '@/app/_components/SignInModal';
+import HeaderSignInModal from '@/app/_components/HeaderSignInModal';
 import { Popover, PopoverContent, PopoverTrigger } from '@/components/ui/popover';
 import { LocationIcon } from '@/components/ui/location-icon';
-import { useTelegramMiniAppEnv } from '@/hooks/use-telegram-mini-app-env';
-import { toast } from '@/hooks/use-toast';
-import { useTelegramAuth } from '@/hooks/use-telegram-auth';
-import { ApiError } from '@/lib/api-errors';
-import { subscribeTelegramSignInRequest } from '@/lib/telegram-auth';
-import { getTelegramAuthBotUsername } from '@/lib/telegram-auth-env';
 import { normalizeLocationTitle } from '@/lib/utils';
-import type { TelegramWidgetUser } from '@/types/telegram-login';
+import HeaderAuthActions from '@/app/_components/HeaderAuthActions';
 
 export interface HeaderActionsProps {
   readonly country: string;
@@ -29,72 +22,9 @@ export default function HeaderActions(props: HeaderActionsProps) {
   const githubUrl = process.env.NEXT_PUBLIC_GITHUB_URL;
   const suggestGigUrl = showSuggestGig ? process.env.NEXT_PUBLIC_SUGGEST_GIG_LINK : undefined;
 
-  const { authState, signIn, signOut } = useTelegramAuth();
-
   const [mobileMenuOpen, setMobileMenuOpen] = useState(false);
   const [desktopMenuOpen, setDesktopMenuOpen] = useState(false);
   const [locationTipOpen, setLocationTipOpen] = useState(false);
-  const [signInModalOpen, setSignInModalOpen] = useState(false);
-
-  const telegramBotUsername = getTelegramAuthBotUsername();
-  const miniAppEnv = useTelegramMiniAppEnv();
-
-  const closeMenus = useCallback(() => {
-    setDesktopMenuOpen(false);
-    setMobileMenuOpen(false);
-  }, []);
-
-  const handleAuthenticated = useCallback(
-    async (user: TelegramWidgetUser) => {
-      try {
-        const { profile } = await signIn(user);
-        const label =
-          profile.displayLabel || (user.username ? `@${user.username}` : user.first_name);
-        toast({
-          title: 'Signed in',
-          description: label,
-        });
-      } catch (e) {
-        const description =
-          e instanceof ApiError ? e.message : 'Could not complete sign in. Please try again.';
-        toast({
-          title: 'Sign in failed',
-          description,
-          variant: 'destructive',
-        });
-      }
-    },
-    [signIn],
-  );
-
-  const openSignInModal = useCallback(() => {
-    closeMenus();
-    setSignInModalOpen(true);
-  }, [closeMenus]);
-
-  useEffect(() => {
-    return subscribeTelegramSignInRequest(() => {
-      openSignInModal();
-    });
-  }, [openSignInModal]);
-
-  const handleSignOut = useCallback(async () => {
-    const label = authState?.displayLabel;
-    await signOut();
-    closeMenus();
-    toast({
-      title: 'Signed out',
-      ...(label ? { description: label } : {}),
-    });
-  }, [authState, signOut, closeMenus]);
-
-  const authMenuProps = {
-    telegramBotUsername,
-    authState,
-    miniAppEnv,
-    onSignInClick: openSignInModal,
-    onSignOut: handleSignOut,
-  };
 
   return (
     <div className="min-w-0 justify-self-end flex items-center space-x-4">
@@ -177,7 +107,7 @@ export default function HeaderActions(props: HeaderActionsProps) {
             align="end"
             className="w-auto min-w-[13rem] max-w-[min(100vw-2rem,20rem)] p-2"
           >
-            <HeaderAuthActions {...authMenuProps} />
+            <HeaderAuthActions />
             <a
               href="/about"
               className="flex items-center gap-2 rounded-md px-2 py-2 text-sm hover:bg-muted"
@@ -207,7 +137,7 @@ export default function HeaderActions(props: HeaderActionsProps) {
             className="w-auto min-w-[14rem] max-w-[min(100vw-2rem,20rem)] p-2"
           >
             <div className="flex flex-col gap-1">
-              <HeaderAuthActions {...authMenuProps} />
+              <HeaderAuthActions />
 
               {!!suggestGigUrl && (
                 <>
@@ -289,13 +219,7 @@ export default function HeaderActions(props: HeaderActionsProps) {
           </PopoverContent>
         </Popover>
       </div>
-
-      <SignInModal
-        isOpen={signInModalOpen}
-        onOpenChange={setSignInModalOpen}
-        telegramBotUsername={telegramBotUsername}
-        onAuthenticated={handleAuthenticated}
-      />
+      <HeaderSignInModal />
     </div>
   );
 }
