+++ src/app/_components/HeaderAuthActions.tsx
@@ -1,22 +1,29 @@
 'use client';
 
 import { LogIn, LogOut } from 'lucide-react';
-import type { TelegramMiniAppEnv } from '@/hooks/use-telegram-mini-app-env';
-import type { TelegramAuthState } from '@/types/telegram-auth';
+import { useTelegramMiniAppEnv } from '@/hooks/use-telegram-mini-app-env';
+import { useTelegramAuth } from '@/hooks/use-telegram-auth';
+import { getTelegramAuthBotUsername } from '@/lib/telegram-auth-env';
+import { useCallback } from 'react';
+import { toast } from '@/hooks/use-toast';
+import { requestTelegramSignIn } from '@/lib/telegram-auth';
 
 const menuRowClass =
   'flex w-full items-center gap-2 rounded-md px-2 py-2 text-left text-sm hover:bg-muted';
 
-export interface HeaderAuthActionsProps {
-  readonly telegramBotUsername?: string;
-  readonly authState: TelegramAuthState | null;
-  readonly miniAppEnv: TelegramMiniAppEnv;
-  readonly onSignInClick: () => void;
-  readonly onSignOut: () => void | Promise<void>;
-}
+export default function HeaderAuthActions() {
+  const { authState, signOut } = useTelegramAuth();
+  const telegramBotUsername = getTelegramAuthBotUsername();
+  const miniAppEnv = useTelegramMiniAppEnv();
 
-export default function HeaderAuthActions(props: HeaderAuthActionsProps) {
-  const { telegramBotUsername, authState, miniAppEnv, onSignInClick, onSignOut } = props;
+  const handleSignOut = useCallback(async () => {
+    const label = authState?.displayLabel;
+    await signOut();
+    toast({
+      title: 'Signed out',
+      ...(label ? { description: label } : {}),
+    });
+  }, [authState, signOut]);
 
   if (!telegramBotUsername?.trim()) {
     return null;
@@ -32,6 +39,7 @@ export default function HeaderAuthActions(props: HeaderAuthActionsProps) {
           <div className="flex min-w-0 flex-1 cursor-default items-center gap-2">
             {authState.photoUrl ? (
               <span className="relative inline-flex h-8 w-8 shrink-0 overflow-hidden rounded-full bg-muted">
+                {/* eslint-disable-next-line @next/next/no-img-element */}
                 <img
                   src={authState.photoUrl}
                   alt=""
@@ -52,7 +60,7 @@ export default function HeaderAuthActions(props: HeaderAuthActionsProps) {
               type="button"
               className="shrink-0 rounded-md p-1.5 text-muted-foreground transition-colors hover:bg-background/80 hover:text-foreground"
               onClick={() => {
-                void onSignOut();
+                void handleSignOut();
               }}
               aria-label="Sign out"
               title="Sign out"
@@ -68,7 +76,7 @@ export default function HeaderAuthActions(props: HeaderAuthActionsProps) {
         <button
           type="button"
           className={menuRowClass}
-          onClick={onSignInClick}
+          onClick={() => requestTelegramSignIn()}
           aria-label="Sign in"
         >
           <LogIn
